+++ Marlin/src/lcd/extui/lib/dgus_creality/creality_touch/AxisSettingsHandler.h
@@ -8,7 +8,7 @@ class AxisSettingsHandler {
         static uint16_t axis_settings_title_icon;
 
         static float axis_steps_mm;
-        static uint32_t max_acceleration_mm_per_s2;
+        static uint16_t max_acceleration_mm_per_s2;
 
         static float jerk;
         static feedRate_t max_feedrate;
@@ -16,7 +16,7 @@ class AxisSettingsHandler {
         static bool has_tmc_settings;
         static uint16_t tmc_current;
         static bool stealthchop;
-        static uint32_t hybrid_threshold;
+        static uint16_t hybrid_threshold;
 
     public:
         static void HandleNavigation(DGUS_VP_Variable &var, void *val_ptr);
